+++ PunishmentManagerCore/src/main/java/me/superbiebel/punishmentmanager/offenseprocessing/abstraction/OffenseProcessorFactory.java
@@ -1,4 +1,4 @@
-package me.superbiebel.punishmentmanager.offenseprocessing;
+package me.superbiebel.punishmentmanager.offenseprocessing.abstraction;
 
 public interface OffenseProcessorFactory {
     public OffenseProcessor giveOffenseExecutor() throws Exception;
